Simplify TestPXEBoot
